refactor(issue): split 839-line issue.js by feature group
issue.js shrinks to a router that wires the issue command to three
feature modules:

  bin/commands/issue/
  ├── crud.js     list / view / create / edit / delete
  ├── comment.js  comment add / list / edit / delete
  ├── link.js     remote-link list / add / update / delete
  └── shared.js   readDescriptionFile (used by crud and comment)

No behavior change — same CLI surface, same flags, same help text. All
251 tests pass and `jira issue --help` / `comment --help` /
`remote-link --help` render identically.
