+++ bin/commands/issue.js
@@ -1,839 +1,18 @@
 const { Command } = require('commander');
-const {
-  createIssuesTable,
-  displayIssueDetails,
-  formatIssueAsMarkdown,
-  buildJQL,
-  createCommentsTable,
-  createRemoteLinksTable
-} = require('../../lib/utils');
-const chalk = require('chalk');
-const fs = require('fs');
-const path = require('path');
+const crud = require('./issue/crud');
+const comment = require('./issue/comment');
+const link = require('./issue/link');
 
 function createIssueCommand(factory) {
   const command = new Command('issue')
     .description('Manage JIRA issues')
     .alias('i');
 
-  // Add subcommands
-  command
-    .command('list')
-    .description('list issues with advanced filtering\n\n' +
-      'Examples:\n' +
-      '  $ jira issue list                              # List recent issues\n' +
-      '  $ jira issue list --assignee=currentUser      # Your assigned issues\n' +
-      '  $ jira issue list --status=Open --limit=50    # Open issues (max 50)\n' +
-      '  $ jira issue list --project=TEST --type=Bug   # Bugs in TEST project')
-    .alias('ls')
-    .option('--project <project>', 'filter by project key')
-    .option('--assignee <assignee>', 'filter by assignee (use "currentUser" for yourself)')
-    .option('--status <status>', 'filter by status (e.g., Open, In Progress)')
-    .option('--type <type>', 'filter by issue type (e.g., Bug, Story)')
-    .option('--reporter <reporter>', 'filter by reporter (use "currentUser" for yourself)')
-    .option('--priority <priority>', 'filter by priority (e.g., High, Medium)')
-    .option('--created <date>', 'created since date (e.g., -7d, 2023-01-01)')
-    .option('--updated <date>', 'updated since date (e.g., -7d, 2023-01-01)')
-    .option('--limit <limit>', 'limit number of results (default: 20)', '20')
-    .option('--jql <query>', 'custom JQL expression, AND-composed with other filters')
-    .action(async (options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-
-      try {
-        await listIssues(client, io, options);
-      } catch (err) {
-        io.error(`Failed to list issues: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  command
-    .command('view <key>')
-    .description('view issue details\n\n' +
-      'Examples:\n' +
-      '  $ jira issue view PROJ-123                           # View in terminal\n' +
-      '  $ jira issue view PROJ-123 --format markdown         # View as markdown\n' +
-      '  $ jira issue view PROJ-123 --output ./issue.md       # Save to file\n' +
-      '  $ jira issue view PROJ-123 --format markdown --output ./issue.md')
-    .alias('show')
-    .option('--format <format>', 'output format (terminal, markdown)', 'terminal')
-    .option('--output <path>', 'save to file instead of displaying')
-    .action(async (key, options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-
-      try {
-        await getIssue(client, io, key, options);
-      } catch (err) {
-        io.error(`Failed to get issue: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  command
-    .command('create')
-    .description('create a new issue\n\n' +
-      'Examples:\n' +
-      '  $ jira issue create --project=TEST --type=Bug --summary="Login fails"\n' +
-      '  $ jira issue create --project=TEST --type=Bug \\\n' +
-      '                      --summary="Login fails on Safari"\n' +
-      '  $ jira issue create --project=PROJ --type=Story \\\n' +
-      '                      --summary="Add user profile page" \\\n' +
-      '                      --description="Users need a profile page to manage settings" \\\n' +
-      '                      --assignee=john.doe --priority=High\n' +
-      '  $ jira issue create --project=PROJ --type=Story \\\n' +
-      '                      --summary="Add feature" \\\n' +
-      '                      --description-file=./feature-spec.md')
-    .alias('new')
-    .option('--project <project>', 'project key (e.g., TEST, PROJ)')
-    .option('--type <type>', 'issue type (e.g., Bug, Story, Task)')
-    .option('--summary <summary>', 'issue summary (required)')
-    .option('--description <description>', 'issue description (optional)')
-    .option('--description-file <path>', 'read description from file (optional)')
-    .option('--assignee <assignee>', 'assignee username')
-    .option('--priority <priority>', 'priority (e.g., High, Medium, Low)')
-    .action(async (options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-      
-      try {
-        await createIssue(client, io, factory, options);
-      } catch (err) {
-        io.error(`Failed to create issue: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  command
-    .command('edit <key>')
-    .description('edit issue')
-    .alias('update')
-    .option('--summary <summary>', 'new summary')
-    .option('--description <description>', 'new description')
-    .option('--description-file <path>', 'read description from file')
-    .option('--assignee <assignee>', 'new assignee')
-    .option('--priority <priority>', 'new priority')
-    .action(async (key, options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-      
-      try {
-        await updateIssue(client, io, key, options);
-      } catch (err) {
-        io.error(`Failed to update issue: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  command
-    .command('delete <key>')
-    .description('delete issue')
-    .alias('rm')
-    .option('-f, --force', 'force delete without confirmation')
-    .action(async (key, options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-      
-      try {
-        await deleteIssue(client, io, key, options);
-      } catch (err) {
-        io.error(`Failed to delete issue: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  // Comment subcommand
-  const commentCommand = command
-    .command('comment')
-    .description('Manage issue comments')
-    .alias('c');
-
-  commentCommand
-    .command('add <key> [text]')
-    .description('add a comment to an issue\n\n' +
-      'Examples:\n' +
-      '  $ jira issue comment add PROJ-123 "Review completed"\n' +
-      '  $ jira issue comment add PROJ-123 --file ./notes.md\n' +
-      '  $ jira issue comment add PROJ-123 "Internal note" --internal')
-    .option('--file <path>', 'read comment body from file')
-    .option('--internal', 'mark comment as internal/private')
-    .action(async (key, text, options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-      
-      try {
-        await addComment(client, io, key, text, options);
-      } catch (err) {
-        io.error(`Failed to add comment: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  commentCommand
-    .command('list <key>')
-    .description('list comments on an issue\n\n' +
-      'Examples:\n' +
-      '  $ jira issue comment list PROJ-123\n' +
-      '  $ jira issue comment list PROJ-123 --format json')
-    .option('--format <format>', 'output format (table, json)', 'table')
-    .action(async (key, options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-      
-      try {
-        await listComments(client, io, key, options);
-      } catch (err) {
-        io.error(`Failed to list comments: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  commentCommand
-    .command('edit <commentId> [text]')
-    .description('edit an existing comment\n\n' +
-      'Examples:\n' +
-      '  $ jira issue comment edit 12345 "Updated comment"\n' +
-      '  $ jira issue comment edit 12345 --file ./updated-notes.md')
-    .option('--file <path>', 'read comment body from file')
-    .action(async (commentId, text, options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-      
-      try {
-        await editComment(client, io, commentId, text, options);
-      } catch (err) {
-        io.error(`Failed to edit comment: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  commentCommand
-    .command('delete <commentId>')
-    .description('delete a comment\n\n' +
-      'Examples:\n' +
-      '  $ jira issue comment delete 12345 --force')
-    .option('-f, --force', 'force delete without confirmation')
-    .action(async (commentId, options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-
-      try {
-        await deleteComment(client, io, commentId, options);
-      } catch (err) {
-        io.error(`Failed to delete comment: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  // Remote link subcommand
-  const remoteLinkCommand = command
-    .command('remote-link')
-    .description('Manage issue remote links')
-    .alias('rl');
-
-  remoteLinkCommand
-    .command('list <key>')
-    .description('list remote links on an issue\n\n' +
-      'Examples:\n' +
-      '  $ jira issue remote-link list PROJ-123\n' +
-      '  $ jira issue remote-link list PROJ-123 --format json\n' +
-      '  $ jira issue remote-link list PROJ-123 --global-id https://example.com/resource')
-    .option('--format <format>', 'output format (table, json)', 'table')
-    .option('--global-id <id>', 'filter by globalId')
-    .action(async (key, options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-
-      try {
-        await listRemoteLinks(client, io, key, options);
-      } catch (err) {
-        io.error(`Failed to list remote links: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  remoteLinkCommand
-    .command('add <key>')
-    .description('add a remote link to an issue\n\n' +
-      'Examples:\n' +
-      '  $ jira issue remote-link add PROJ-123 \\\n' +
-      '      --url https://github.com/org/repo/pull/42 \\\n' +
-      '      --title "org/repo#42"\n' +
-      '  $ jira issue remote-link add PROJ-123 \\\n' +
-      '      --url https://github.com/org/repo/pull/42 \\\n' +
-      '      --title "org/repo#42" \\\n' +
-      '      --global-id https://github.com/org/repo/pull/42 \\\n' +
-      '      --relationship "relates to"')
-    .option('--url <url>', 'external resource URL (required)')
-    .option('--title <title>', 'link title (required)')
-    .option('--global-id <id>', 'stable identifier for upsert (recommended)')
-    .option('--relationship <relationship>', 'link relationship (e.g., "relates to")')
-    .option('--summary <summary>', 'optional summary shown under the title')
-    .option('--icon-url <url>', 'optional 16x16 icon URL')
-    .option('--icon-title <title>', 'optional icon alt text')
-    .action(async (key, options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-
-      try {
-        await addRemoteLink(client, io, key, options);
-      } catch (err) {
-        io.error(`Failed to add remote link: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  remoteLinkCommand
-    .command('update <key> <linkId>')
-    .description('update an existing remote link\n\n' +
-      'Examples:\n' +
-      '  $ jira issue remote-link update PROJ-123 12345 --title "Updated title"\n' +
-      '  $ jira issue remote-link update PROJ-123 12345 --url https://example.com/new')
-    .option('--url <url>', 'new resource URL')
-    .option('--title <title>', 'new link title')
-    .option('--relationship <relationship>', 'new relationship')
-    .option('--summary <summary>', 'new summary')
-    .option('--icon-url <url>', 'new icon URL')
-    .option('--icon-title <title>', 'new icon alt text')
-    .action(async (key, linkId, options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-
-      try {
-        await updateRemoteLink(client, io, key, linkId, options);
-      } catch (err) {
-        io.error(`Failed to update remote link: ${err.message}`);
-        process.exit(1);
-      }
-    });
-
-  remoteLinkCommand
-    .command('delete <key> <linkId>')
-    .description('delete a remote link\n\n' +
-      'Examples:\n' +
-      '  $ jira issue remote-link delete PROJ-123 12345 --force')
-    .option('-f, --force', 'force delete without confirmation')
-    .action(async (key, linkId, options) => {
-      const io = factory.getIOStreams();
-      const client = await factory.getJiraClient();
-
-      try {
-        await deleteRemoteLink(client, io, key, linkId, options);
-      } catch (err) {
-        io.error(`Failed to delete remote link: ${err.message}`);
-        process.exit(1);
-      }
-    });
+  crud.register(command, factory);
+  comment.register(command, factory);
+  link.register(command, factory);
 
   return command;
-}
-
-function readDescriptionFile(filePath) {
-  const absolutePath = path.resolve(filePath);
-
-  if (!fs.existsSync(absolutePath)) {
-    throw new Error(`Description file not found: ${absolutePath}`);
-  }
-
-  const stats = fs.statSync(absolutePath);
-  if (!stats.isFile()) {
-    throw new Error(`Path is not a file: ${absolutePath}`);
-  }
-
-  const content = fs.readFileSync(absolutePath, 'utf8');
-  if (!content.trim()) {
-    throw new Error(`Description file is empty: ${absolutePath}`);
-  }
-
-  return content;
-}
-
-async function listIssues(client, io, options) {
-  const spinner = io.spinner('Fetching issues...');
-  
-  try {
-    const jql = buildJQL(options);
-    const limit = parseInt(options.limit) || 20;
-    const result = await client.searchIssues(jql, {
-      maxResults: limit
-    });
-
-    spinner.stop();
-
-    if (result.issues.length === 0) {
-      io.info('No issues found');
-      return;
-    }
-
-    io.out(chalk.bold(`\nFound ${result.issues.length} issues (showing ${Math.min(result.issues.length, limit)}):\n`));
-    
-    const table = createIssuesTable(result.issues);
-    io.out(table.toString());
-
-    if (result.total > result.issues.length) {
-      io.info(`Showing ${result.issues.length} of ${result.total} total issues`);
-    }
-
-  } catch (err) {
-    spinner.stop();
-    throw err;
-  }
-}
-
-async function getIssue(client, io, issueKey, options = {}) {
-  const spinner = io.spinner(`Fetching issue ${issueKey}...`);
-
-  try {
-    const issue = await client.getIssue(issueKey);
-    spinner.stop();
-
-    if (options.output) {
-      const outputPath = path.resolve(options.output);
-      const content = formatIssueAsMarkdown(issue);
-
-      fs.writeFileSync(outputPath, content, 'utf8');
-      io.success(`Issue ${issueKey} saved to ${outputPath}`);
-    } else if (options.format === 'markdown') {
-      const markdown = formatIssueAsMarkdown(issue);
-      io.out('\n' + markdown);
-    } else {
-      displayIssueDetails(issue, io);
-    }
-
-  } catch (err) {
-    spinner.stop();
-    throw err;
-  }
-}
-
-async function createIssue(client, io, factory, options = {}) {
-  const config = factory.getConfig();
-
-  // Validate required options
-  if (!options.project || !options.type || !options.summary) {
-    throw new Error(
-      'Missing required options for creating an issue.\n\n' +
-      'Usage: jira issue create --project <KEY> --type <TYPE> --summary <TEXT>\n\n' +
-      'Required options:\n' +
-      '  --project <KEY>         Project key (e.g., TEST, PROJ)\n' +
-      '  --type <TYPE>           Issue type (e.g., Bug, Story, Task)\n' +
-      '  --summary <TEXT>        Issue summary\n\n' +
-      'Optional:\n' +
-      '  --description <TEXT>    Issue description\n' +
-      '  --description-file <PATH>  Read description from file\n' +
-      '  --assignee <USER>       Assignee username\n' +
-      '  --priority <PRIORITY>   Priority level\n\n' +
-      'Example:\n' +
-      '  jira issue create --project TEST --type Bug --summary "Login fails"'
-    );
-  }
-
-  // Handle description from file or text
-  let description = '';
-  if (options.description && options.descriptionFile) {
-    throw new Error('Cannot use both --description and --description-file. Please use only one.');
-  }
-
-  if (options.descriptionFile) {
-    description = readDescriptionFile(options.descriptionFile);
-  } else if (options.description) {
-    description = options.description;
-  }
-
-  const issueData = {
-    fields: {
-      project: { key: options.project },
-      issuetype: { name: options.type },
-      summary: options.summary,
-      description: description
-    }
-  };
-
-  if (options.assignee) {
-    issueData.fields.assignee = { name: options.assignee };
-  }
-
-  if (options.priority) {
-    issueData.fields.priority = { name: options.priority };
-  }
-
-  const spinner = io.spinner('Creating issue...');
-  const result = await client.createIssue(issueData);
-  spinner.stop();
-
-  io.success(`Issue created: ${result.key}`);
-  io.out(`URL: ${config.get('server')}/browse/${result.key}`);
-}
-
-async function updateIssue(client, io, issueKey, options = {}) {
-  // Validate at least one option provided
-  if (!options.summary && !options.description && !options.descriptionFile &&
-      !options.assignee && !options.priority) {
-    throw new Error(
-      'At least one field must be specified for update.\n\n' +
-      'Usage: jira issue edit <KEY> [options]\n\n' +
-      'Available options:\n' +
-      '  --summary <TEXT>          New summary\n' +
-      '  --description <TEXT>      New description\n' +
-      '  --description-file <PATH> Read description from file\n' +
-      '  --assignee <USER>         New assignee\n' +
-      '  --priority <PRIORITY>     New priority\n\n' +
-      'Example:\n' +
-      '  jira issue edit PROJ-123 --summary "Updated summary"\n' +
-      '  jira issue edit PROJ-123 --description-file ./updated-spec.md'
-    );
-  }
-
-  // Get current issue
-  const spinner = io.spinner(`Loading issue ${issueKey}...`);
-  const issue = await client.getIssue(issueKey);
-  spinner.stop();
-
-  io.out(chalk.bold(`\nUpdating issue: ${issue.key}`));
-  io.out(`Current summary: ${issue.fields.summary}\n`);
-
-  const updateData = { fields: {} };
-  let hasChanges = false;
-
-  if (options.summary && options.summary !== issue.fields.summary) {
-    updateData.fields.summary = options.summary;
-    hasChanges = true;
-  }
-
-  // Handle description from file or text
-  if (options.description && options.descriptionFile) {
-    throw new Error('Cannot use both --description and --description-file. Please use only one.');
-  }
-
-  if (options.descriptionFile) {
-    const description = readDescriptionFile(options.descriptionFile);
-    if (description !== (issue.fields.description || '')) {
-      updateData.fields.description = description;
-      hasChanges = true;
-    }
-  } else if (options.description && options.description !== (issue.fields.description || '')) {
-    updateData.fields.description = options.description;
-    hasChanges = true;
-  }
-
-  if (options.assignee) {
-    updateData.fields.assignee = { name: options.assignee };
-    hasChanges = true;
-  }
-
-  if (options.priority) {
-    updateData.fields.priority = { name: options.priority };
-    hasChanges = true;
-  }
-
-  if (!hasChanges) {
-    io.info('No changes made');
-    return;
-  }
-
-  const updateSpinner = io.spinner('Updating issue...');
-  await client.updateIssue(issueKey, updateData);
-  updateSpinner.stop();
-
-  io.success(`Issue ${issueKey} updated successfully`);
-}
-
-async function deleteIssue(client, io, issueKey, options = {}) {
-  // Get issue details first
-  const spinner = io.spinner(`Loading issue ${issueKey}...`);
-  const issue = await client.getIssue(issueKey);
-  spinner.stop();
-
-  io.out(chalk.bold.red('\nWARNING: You are about to delete this issue:'));
-  io.out(`  Key: ${chalk.cyan(issue.key)}`);
-  io.out(`  Summary: ${issue.fields.summary}`);
-  io.out(`  Type: ${issue.fields.issuetype.name}\n`);
-
-  if (!options.force) {
-    throw new Error(
-      'Deletion requires --force flag to confirm.\n' +
-      `Use: jira issue delete ${issueKey} --force`
-    );
-  }
-
-  const deleteSpinner = io.spinner('Deleting issue...');
-  await client.deleteIssue(issueKey);
-  deleteSpinner.stop();
-
-  io.success(`Issue ${issueKey} deleted successfully`);
-}
-
-async function addComment(client, io, issueKey, text, options = {}) {
-  // Validate that we have comment text from either argument or file
-  if (!text && !options.file) {
-    throw new Error(
-      'Comment text is required.\n\n' +
-      'Usage: jira issue comment add <KEY> [text] [options]\n\n' +
-      'Provide comment text either as:\n' +
-      '  - Direct argument: jira issue comment add PROJ-123 "Comment text"\n' +
-      '  - From file: jira issue comment add PROJ-123 --file ./comment.md\n\n' +
-      'Options:\n' +
-      '  --file <path>  Read comment body from file\n' +
-      '  --internal     Mark comment as internal/private'
-    );
-  }
-
-  if (text && options.file) {
-    throw new Error('Cannot use both text argument and --file option. Please use only one.');
-  }
-
-  let commentBody = text;
-  if (options.file) {
-    commentBody = readDescriptionFile(options.file);
-  }
-
-  const spinner = io.spinner(`Adding comment to ${issueKey}...`);
-  const comment = await client.addComment(issueKey, commentBody, {
-    internal: options.internal
-  });
-  spinner.stop();
-
-  io.success(`Comment added to ${issueKey}`);
-  io.out(`Comment ID: ${comment.id}`);
-  
-  if (options.internal) {
-    io.info('Comment marked as internal');
-  }
-}
-
-async function listComments(client, io, issueKey, options = {}) {
-  const spinner = io.spinner(`Fetching comments for ${issueKey}...`);
-  
-  try {
-    const result = await client.getComments(issueKey);
-    spinner.stop();
-
-    if (!result.comments || result.comments.length === 0) {
-      io.info(`No comments found for ${issueKey}`);
-      return;
-    }
-
-    io.out(chalk.bold(`\nComments for ${issueKey} (${result.comments.length} total):\n`));
-
-    if (options.format === 'json') {
-      io.out(JSON.stringify(result.comments, null, 2));
-    } else {
-      const table = createCommentsTable(result.comments);
-      io.out(table.toString());
-    }
-
-  } catch (err) {
-    spinner.stop();
-    throw err;
-  }
-}
-
-async function editComment(client, io, commentId, text, options = {}) {
-  // Validate that we have comment text from either argument or file
-  if (!text && !options.file) {
-    throw new Error(
-      'Comment text is required.\n\n' +
-      'Usage: jira issue comment edit <COMMENT-ID> [text] [options]\n\n' +
-      'Provide comment text either as:\n' +
-      '  - Direct argument: jira issue comment edit 12345 "Updated text"\n' +
-      '  - From file: jira issue comment edit 12345 --file ./updated.md\n\n' +
-      'Options:\n' +
-      '  --file <path>  Read comment body from file'
-    );
-  }
-
-  if (text && options.file) {
-    throw new Error('Cannot use both text argument and --file option. Please use only one.');
-  }
-
-  let commentBody = text;
-  if (options.file) {
-    commentBody = readDescriptionFile(options.file);
-  }
-
-  const spinner = io.spinner(`Updating comment ${commentId}...`);
-  await client.updateComment(commentId, commentBody);
-  spinner.stop();
-
-  io.success(`Comment ${commentId} updated successfully`);
-}
-
-async function deleteComment(client, io, commentId, options = {}) {
-  io.out(chalk.bold.red('\nWARNING: You are about to delete this comment:'));
-  io.out(`  Comment ID: ${chalk.cyan(commentId)}\n`);
-
-  if (!options.force) {
-    throw new Error(
-      'Deletion requires --force flag to confirm.\n' +
-      `Use: jira issue comment delete ${commentId} --force`
-    );
-  }
-
-  const spinner = io.spinner('Deleting comment...');
-  await client.deleteComment(commentId);
-  spinner.stop();
-
-  io.success(`Comment ${commentId} deleted successfully`);
-}
-
-function buildRemoteLinkPayload(options) {
-  const payload = { object: {} };
-
-  if (options.url) payload.object.url = options.url;
-  if (options.title) payload.object.title = options.title;
-  if (options.summary) payload.object.summary = options.summary;
-
-  if (options.iconUrl || options.iconTitle) {
-    payload.object.icon = {};
-    if (options.iconUrl) payload.object.icon.url16x16 = options.iconUrl;
-    if (options.iconTitle) payload.object.icon.title = options.iconTitle;
-  }
-
-  if (options.globalId) payload.globalId = options.globalId;
-  if (options.relationship) payload.relationship = options.relationship;
-
-  return payload;
-}
-
-async function listRemoteLinks(client, io, issueKey, options = {}) {
-  const spinner = io.spinner(`Fetching remote links for ${issueKey}...`);
-
-  try {
-    const links = await client.getRemoteLinks(issueKey, {
-      globalId: options.globalId
-    });
-    spinner.stop();
-
-    if (links.length === 0) {
-      io.info(`No remote links found for ${issueKey}`);
-      return;
-    }
-
-    io.out(chalk.bold(`\nRemote links for ${issueKey} (${links.length} total):\n`));
-
-    if (options.format === 'json') {
-      io.out(JSON.stringify(links, null, 2));
-    } else {
-      const table = createRemoteLinksTable(links);
-      io.out(table.toString());
-    }
-
-  } catch (err) {
-    spinner.stop();
-    throw err;
-  }
-}
-
-async function addRemoteLink(client, io, issueKey, options = {}) {
-  if (!options.url || !options.title) {
-    throw new Error(
-      'Missing required options for adding a remote link.\n\n' +
-      'Usage: jira issue remote-link add <KEY> --url <URL> --title <TITLE> [options]\n\n' +
-      'Required options:\n' +
-      '  --url <URL>              External resource URL\n' +
-      '  --title <TITLE>          Link title\n\n' +
-      'Optional:\n' +
-      '  --global-id <ID>         Stable identifier for upsert behavior\n' +
-      '  --relationship <REL>     Link relationship (e.g., "relates to")\n' +
-      '  --summary <TEXT>         Summary shown under the title\n' +
-      '  --icon-url <URL>         16x16 icon URL\n' +
-      '  --icon-title <TITLE>     Icon alt text\n\n' +
-      'Example:\n' +
-      '  jira issue remote-link add PROJ-123 \\\n' +
-      '      --url https://github.com/org/repo/pull/42 \\\n' +
-      '      --title "org/repo#42" \\\n' +
-      '      --global-id https://github.com/org/repo/pull/42'
-    );
-  }
-
-  const payload = buildRemoteLinkPayload(options);
-
-  const spinner = io.spinner(`Adding remote link to ${issueKey}...`);
-  const result = await client.addRemoteLink(issueKey, payload);
-  spinner.stop();
-
-  io.success(`Remote link added to ${issueKey}`);
-  if (result && result.id) {
-    io.out(`Link ID: ${result.id}`);
-  }
-  if (result && result.self) {
-    io.out(`URL: ${result.self}`);
-  }
-}
-
-function mergeRemoteLinkPayload(existing, options) {
-  const base = { ...(existing || {}) };
-  delete base.id;
-  delete base.self;
-
-  const payload = {
-    ...base,
-    object: { ...(base.object || {}) }
-  };
-
-  if (options.url) payload.object.url = options.url;
-  if (options.title) payload.object.title = options.title;
-  if (options.summary) payload.object.summary = options.summary;
-  if (options.relationship) payload.relationship = options.relationship;
-
-  if (options.iconUrl || options.iconTitle) {
-    payload.object.icon = { ...(payload.object.icon || {}) };
-    if (options.iconUrl) payload.object.icon.url16x16 = options.iconUrl;
-    if (options.iconTitle) payload.object.icon.title = options.iconTitle;
-  }
-
-  return payload;
-}
-
-async function updateRemoteLink(client, io, issueKey, linkId, options = {}) {
-  if (!options.url && !options.title && !options.summary &&
-      !options.relationship && !options.iconUrl && !options.iconTitle) {
-    throw new Error(
-      'At least one field must be specified for update.\n\n' +
-      'Usage: jira issue remote-link update <KEY> <LINK-ID> [options]\n\n' +
-      'Available options:\n' +
-      '  --url <URL>              New resource URL\n' +
-      '  --title <TITLE>          New link title\n' +
-      '  --relationship <REL>     New relationship\n' +
-      '  --summary <TEXT>         New summary\n' +
-      '  --icon-url <URL>         New icon URL\n' +
-      '  --icon-title <TITLE>     New icon alt text\n\n' +
-      'Example:\n' +
-      '  jira issue remote-link update PROJ-123 12345 --title "Updated title"'
-    );
-  }
-
-  const fetchSpinner = io.spinner(`Loading remote link ${linkId}...`);
-  const existing = await client.getRemoteLink(issueKey, linkId);
-  fetchSpinner.stop();
-
-  const payload = mergeRemoteLinkPayload(existing, options);
-
-  const spinner = io.spinner(`Updating remote link ${linkId}...`);
-  await client.updateRemoteLink(issueKey, linkId, payload);
-  spinner.stop();
-
-  io.success(`Remote link ${linkId} updated successfully`);
-}
-
-async function deleteRemoteLink(client, io, issueKey, linkId, options = {}) {
-  io.out(chalk.bold.red('\nWARNING: You are about to delete this remote link:'));
-  io.out(`  Issue: ${chalk.cyan(issueKey)}`);
-  io.out(`  Link ID: ${chalk.cyan(linkId)}\n`);
-
-  if (!options.force) {
-    throw new Error(
-      'Deletion requires --force flag to confirm.\n' +
-      `Use: jira issue remote-link delete ${issueKey} ${linkId} --force`
-    );
-  }
-
-  const spinner = io.spinner('Deleting remote link...');
-  await client.deleteRemoteLink(issueKey, linkId);
-  spinner.stop();
-
-  io.success(`Remote link ${linkId} deleted successfully`);
 }
 
 module.exports = createIssueCommand;
